# openapi.yaml
openapi: "3.1.0"
info:
  title: Stellar Wrap API
  version: "1.0.0"
  description: |
    API routes exposed by the Stellar Wrap Next.js application.

    Public endpoints:
    - **GET /api/wrapped** — returns aggregated on-chain statistics for a Stellar address.
    - **GET /api/og** — returns a 1200 × 1200 PNG share card (served from Vercel Edge Runtime).

    Notification endpoints (push + email subscriptions, GDPR data deletion):
    - **POST /api/notifications/subscribe** — store a browser push subscription.
    - **POST /api/notifications/subscribe-email** — start an email subscription (pending confirmation).
    - **GET /api/notifications/confirm-email** — confirm a pending email subscription.
    - **POST /api/notifications/unsubscribe** — unsubscribe by token or by wallet + channel.
    - **POST /api/notifications/dispatch** — triggered by a Vercel cron; fans out due notifications.
    - **GET /api/notifications/preferences/{wallet}** — read a subscription record.
    - **PUT /api/notifications/preferences/{wallet}** — update a subscription record.
    - **DELETE /api/notifications/data/{wallet}** — GDPR data deletion request.
servers:
  - url: https://stellar-wrap.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
security: []
paths:
...
  /api/notifications/confirm-email:
    get:
      operationId: confirmEmail
      summary: Confirm a pending email subscription
      description: |
        Activates a pending email subscription by matching the confirmation
        token against the stored record for the wallet. On success (including
        the already-active case) the caller is redirected to
        `/notifications?confirmed=true`.
      parameters:
        - name: token
          in: query
          required: true
          description: Confirmation token emailed to the subscriber.
          schema:
            type: string
            example: "f3a5c1d9e2b8476fa0b1c2d3e4f5a6b7"
        - name: wallet
          in: query
          required: true
          description: Stellar public key the subscription belongs to.
          schema:
            type: string
            example: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
      responses:
        "302":
          description: Redirect to `/notifications?confirmed=true` on success.
          headers:
            Location:
              description: Redirect target.
              schema:
                type: string
                example: "/notifications?confirmed=true"
        "400":
          description: Missing token or wallet parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Missing token or wallet parameter"
        "404":
          description: No pending subscription or invalid/expired token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                noSubscription:
                  value:
                    error: "No pending email subscription found"
                invalidToken:
                  value:
                    error: "Invalid or expired confirmation token"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
  /api/notifications/unsubscribe:
    post:
      operationId: unsubscribe
      summary: Unsubscribe from notifications
      description: |
        Two modes, mutually exclusive:
        - `{ "token": "..." }` — one-click unsubscribe using the token from the
          footer of a notification email.
        - `{ "walletAddress": "...", "channel": "push" | "email" }` — disable a
          single channel from the preference page.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnsubscribeRequest"
            examples:
              byToken:
                value:
                  token: "f3a5c1d9e2b8476fa0b1c2d3e4f5a6b7"
              byChannel:
                value:
                  walletAddress: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
                  channel: "email"
      responses:
        "200":
          description: Successfully unsubscribed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
              example:
                ok: true
        "400":
          description: Missing token or walletAddress+channel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Provide either token or walletAddress+channel"
        "404":
          description: Token not found or no subscription for the wallet.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                tokenNotFound:
                  value:
                    error: "Token not found"
                noSubscription:
                  value:
                    error: "No subscription found"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
  /api/notifications/dispatch:
    post:
      operationId: dispatchNotifications
      summary: Fan out due notifications to all subscribers
      description: |
        Evaluates which wrap periods have just started and fans out push/email
        notifications to every matching subscriber. Idempotent per
        wallet/channel/period — a dispatch log entry prevents duplicate sends.

        **Cron**: Driven hourly by a Vercel cron job
        (`vercel.json` → `0 * * * *`).

        **Authentication**: Requires `Authorization: Bearer $CRON_SECRET`.
        If `CRON_SECRET` is not configured on the server, the check is skipped.
      security:
        - cronSecretAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                periods:
                  type: array
                  description: Override which periods to dispatch instead of deriving from the current time.
                  items:
                    $ref: "#/components/schemas/WrapPeriod"
            example:
              periods:
                - monthly
      responses:
        "200":
          description: Dispatch completed.
          content:
            application/json:
              schema:
                type: object
                required: [ok, dispatched, periods]
                properties:
                  ok:
                    type: boolean
                    example: true
                  dispatched:
                    type: integer
                    minimum: 0
                    description: Number of notifications actually sent.
                    example: 42
                  periods:
                    type: array
                    description: Periods considered by this dispatch.
                    items:
                      $ref: "#/components/schemas/WrapPeriod"
                  message:
                    type: string
                    description: Present when no periods were active.
                    example: "No active periods"
              examples:
                withDispatch:
                  value:
                    ok: true
                    dispatched: 42
                    periods:
                      - monthly
                noActivePeriods:
                  value:
                    ok: true
                    dispatched: 0
                    periods: []
                    message: "No active periods"
        "401":
          description: Missing or invalid cron secret.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Unauthorized"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
  /api/notifications/data/{wallet}:
    delete:
      operationId: deleteNotificationData
      summary: Request deletion of all data for a wallet (GDPR)
      description: |
        Immediately purges push/email PII for the wallet and clears its
        dispatch logs. The record is flagged with a deletion timestamp; a
        confirmation email is sent when an email address was on file
        (non-fatal if sending fails). Dispatch logs are otherwise cleared
        within 30 days per policy.
      parameters:
        - name: wallet
          in: path
          required: true
          description: Stellar public key whose data should be deleted.
          schema:
            type: string
            pattern: "^G[A-Z2-7]{55}$"
            example: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
      responses:
        "200":
          description: Data deleted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
              example:
                ok: true
        "400":
          description: Invalid wallet address.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Invalid wallet address"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
...
components:
  securitySchemes:
    cronSecretAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token matching the `CRON_SECRET` environment variable. Set by the
        Vercel cron job that calls `POST /api/notifications/dispatch`.
  schemas:
    WrapPeriod:
      type: string
      enum: [weekly, monthly, yearly]
      example: monthly
...
    UnsubscribeRequest:
      type: object
      description: One of `token` or `walletAddress` + `channel` (mutually exclusive).
      properties:
        token:
          type: string
          description: Unsubscribe token from a notification email.
          example: "f3a5c1d9e2b8476fa0b1c2d3e4f5a6b7"
        walletAddress:
          type: string
          pattern: "^G[A-Z2-7]{55}$"
          description: Stellar public key.
          example: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
        channel:
          type: string
          enum: [push, email]
          description: Channel to disable on the preference page.
          example: "email"
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          example: true
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string